requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
alphashape==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
attrs==26.1.0
Authlib==1.6.9
azure-core==1.39.0
azure-identity==1.25.3
azure-storage-blob==12.28.0
backoff==2.2.1
beautifulsoup4==4.14.3
boto3==1.42.73
botocore==1.42.73
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
chardet==7.3.0
charset-normalizer==3.4.6
click==8.3.1
click-log==0.4.0
cloudpickle==3.1.2
colorama==0.4.6
Crawl4AI==0.8.6
croniter==6.2.2
cryptography==46.0.5
cssselect==1.4.0
curl_cffi==0.14.0
ddgs==9.11.4
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
email-validator==2.3.0
fake-useragent==2.2.0
fastapi==0.135.1
fastapi-sso==0.16.0
fastuuid==0.14.0
filelock==3.25.2
firecrawl-py==4.19.0
frozenlist==1.8.0
fsspec==2026.2.0
generativeai==0.0.1
google-adk==1.27.2
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.142.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.40.1
google-cloud-bigquery-storage==2.36.2
google-cloud-bigtable==2.35.0
google-cloud-core==2.5.0
google-cloud-dataplex==2.16.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.21.0
google-cloud-logging==3.14.0
google-cloud-monitoring==2.29.1
google-cloud-pubsub==2.36.0
google-cloud-resource-manager==1.16.0
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.63.0
google-cloud-speech==2.37.0
google-cloud-storage==3.10.0
google-cloud-trace==1.18.0
google-crc32c==1.8.0
google-genai==1.68.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
graphviz==0.21
greenlet==3.3.2
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.78.0
grpcio-status==1.78.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.2
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lark==1.3.1
litellm==1.82.6
litellm-enterprise==0.1.35
litellm-proxy-extras==0.4.60
lxml==5.4.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
msal==1.35.1
msal-extensions==1.3.1
multidict==6.7.1
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.3
oauthlib==3.3.1
openai==2.29.0
opentelemetry-api==1.38.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.7
outcome==1.3.0.post0
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049096285/work
pandas==3.0.1
patchright==1.58.2
pillow==12.1.1
playwright==1.58.0
playwright-stealth==2.0.2
pluggy==1.6.0
polars==1.39.3
polars-runtime-32==1.39.3
primp==1.1.3
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
PyJWT==2.12.1
PyNaCl==1.6.2
pyOpenSSL==26.0.0
pyparsing==3.3.2
pypdf==6.9.2
PySocks==1.7.1
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pywin32==311
PyYAML==6.0.3
rank-bm25==0.2.2
redis==7.3.0
referencing==0.37.0
regex==2026.2.28
requests==2.33.0
rich==13.9.4
rpds-py==0.30.0
rq==2.7.0
rtree==1.4.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
selenium==4.41.0
sentence-transformers==5.3.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlalchemy-spanner==1.17.2
sqlparse==0.5.5
sse-starlette==3.3.3
starlette==0.52.1
sympy==1.14.0
tavily-python==0.7.23
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.4.0
trimesh==4.11.5
trio==0.33.0
trio-websocket==0.12.2
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
unclecode-litellm==1.81.13
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
watchdog==6.0.0
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.3.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
pymysql        # Driver để Python nói chuyện được với MySQL